+++ deltafi-ui/CHANGELOG.md
@@ -3,10 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [0.21.0]
 
 ### Added
 - Display Content-level MetaData in Content Dialog (#132)
+- Added Copy to Clipboard button to Content Viewer (#144)
+- Loading indicator for Parent/Child DeltaFiles (#142)
+- Paging for Parent/Child DeltaFiles (#143)
 
 ### Changed
 - Max page size to 1000 on Search and Errors pages (#139)
@@ -156,7 +159,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - NiFi link from Dashboard
 - Requested resource information from System Metrics
 
-[Unreleased]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.20.0...main
+[0.21.0]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.20.0...0.21.0
 [0.20.0]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.19.1...0.20.0
 [0.19.1]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.19.0...0.19.1
 [0.19.0]: https://gitlab.com/systolic/deltafi/deltafi-ui/-/compare/0.18.1...0.19.0